requirements.txt
python-dotenv
pandas
numpy
pymongo
# pymongo[srv]==3.6
certifi
scikit-learn
pyaml
mlflow<3
dagshub
fastapi
uvicorn
python-multipart

# -e .